updatePeriodMillis
Request updates to the widget once an hour; the system gets to decide exactly when the update takes place (and will wake the device). Widgets will receive an `ACTION_APPWIDGET_UPDATE` which should restore AlarmManager based updates (scheduled once a minute and will not wake the device). This may resolve issues where the widget fails to update reliably on some devices (#15).